ENH: poseidon implementation with an initial set of params
The module permutation.rs implements the poseidon permutation
depending on its set of parameters, as well as the corresponding
hash function via the Hades strategy.

The module paramaters and its child modules implement the different
set of parameters. Only one set of starkad parameters for a prime
field is implemented, but more will be added.