refactor: poseidon permutation and sponge
WIP: some tests fail at the moment.

Following the no_std refactoring for fields, this commit starts integrating the new fields to poseidon.
A refactoring was necessary to make it no_std and also cleaner in meaning.
However, no_std means complicated generics, it could not be so easily avoided.