refactor: no_std refactoring with montgomery's form
Implementation of a prime field in montgomery's form is completed.
Poseidon hashing is compatible with montgomery's form. This means it is optimized for multiplications and exponentiation, but at an extra cost of conversion when absorbing or squeezing.
Starkware's sw2 variant is migrated, as constants are better kept in montgomery's form to avoid many conversions.
Cargo.toml has been improved to include variants and std as features.